Inclusão do mapeamento do Cliente ao Context. Alterada a classe CustomValidationAttributeAdapterProvider de lugar para junto da LocalizationConfig
